requirements.txt
ollama
python-dotenv
requests
